the ClearTextButtonBehavior was a good start, but this is better
the IsClearTextButtonBehaviorEnabled sets the button click event if its
set to true
