+++ MahApps.Metro/Controls/Helper/TextBoxHelper.cs
@@ -5,7 +5,6 @@ using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
 using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Interactivity;
 using System.Windows.Media;
 
 namespace MahApps.Metro.Controls
@@ -23,6 +22,10 @@ namespace MahApps.Metro.Controls
         public static readonly DependencyProperty UseFloatingWatermarkProperty = DependencyProperty.RegisterAttached("UseFloatingWatermark", typeof(bool), typeof(TextBoxHelper), new FrameworkPropertyMetadata(false, ButtonCommandOrClearTextChanged));
         public static readonly DependencyProperty TextLengthProperty = DependencyProperty.RegisterAttached("TextLength", typeof(int), typeof(TextBoxHelper), new UIPropertyMetadata(0));
         public static readonly DependencyProperty ClearTextButtonProperty = DependencyProperty.RegisterAttached("ClearTextButton", typeof(bool), typeof(TextBoxHelper), new FrameworkPropertyMetadata(false, ButtonCommandOrClearTextChanged));
+        /// <summary>
+        /// The clear text button behavior property. It sets a click event to the button if the value is true.
+        /// </summary>
+        public static readonly DependencyProperty IsClearTextButtonBehaviorEnabledProperty = DependencyProperty.RegisterAttached("IsClearTextButtonBehaviorEnabled", typeof(bool), typeof(TextBoxHelper), new FrameworkPropertyMetadata(false, IsClearTextButtonBehaviorEnabledChanged));
         
         public static readonly DependencyProperty ButtonCommandProperty = DependencyProperty.RegisterAttached("ButtonCommand", typeof(ICommand), typeof(TextBoxHelper), new FrameworkPropertyMetadata(null, ButtonCommandOrClearTextChanged));
         public static readonly DependencyProperty ButtonCommandParameterProperty = DependencyProperty.RegisterAttached("ButtonCommandParameter", typeof(object), typeof(TextBoxHelper), new FrameworkPropertyMetadata(null));
@@ -316,6 +319,24 @@ namespace MahApps.Metro.Controls
             obj.SetValue(ClearTextButtonProperty, value);
         }
 
+        /// <summary>
+        /// Gets the clear text button behavior.
+        /// </summary>
+        [AttachedPropertyBrowsableForType(typeof(Button))]
+        public static bool GetIsClearTextButtonBehaviorEnabled(Button d)
+        {
+            return (bool)d.GetValue(IsClearTextButtonBehaviorEnabledProperty);
+        }
+
+        /// <summary>
+        /// Sets the clear text button behavior.
+        /// </summary>
+        [AttachedPropertyBrowsableForType(typeof(Button))]
+        public static void SetIsClearTextButtonBehaviorEnabled(Button obj, bool value)
+        {
+            obj.SetValue(IsClearTextButtonBehaviorEnabledProperty, value);
+        }
+
         public static ICommand GetButtonCommand(DependencyObject d)
         {
             return (ICommand)d.GetValue(ButtonCommandProperty);
@@ -364,6 +385,57 @@ namespace MahApps.Metro.Controls
         public static void SetButtonFontFamily(DependencyObject obj, FontFamily value)
         {
             obj.SetValue(ButtonFontFamilyProperty, value);
+        }
+
+        private static void IsClearTextButtonBehaviorEnabledChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
+        {
+            var button = d as Button;
+            if (e.OldValue != e.NewValue && button != null)
+            {
+                button.Click -= ButtonClicked;
+                if ((bool)e.NewValue)
+                {
+                    button.Click += ButtonClicked;
+                }
+            }
+        }
+
+        public static void ButtonClicked(object sender, RoutedEventArgs e)
+        {
+            var button = ((Button)sender);
+            var parent = VisualTreeHelper.GetParent(button);
+            while (!(parent is TextBox || parent is PasswordBox || parent is ComboBox))
+            {
+                parent = VisualTreeHelper.GetParent(parent);
+            }
+
+            var command = GetButtonCommand(parent);
+            if (command != null && command.CanExecute(parent))
+            {
+                var commandParameter = GetButtonCommandParameter(parent);
+
+                command.Execute(commandParameter ?? parent);
+            }
+
+            if (GetClearTextButton(parent))
+            {
+                if (parent is TextBox)
+                {
+                    ((TextBox)parent).Clear();
+                }
+                else if (parent is PasswordBox)
+                {
+                    ((PasswordBox)parent).Clear();
+                }
+                else if (parent is ComboBox)
+                {
+                    if (((ComboBox)parent).IsEditable)
+                    {
+                        ((ComboBox)parent).Text = string.Empty;
+                    }
+                    ((ComboBox)parent).SelectedItem = null;
+                }
+            }
         }
 
         private static void ButtonCommandOrClearTextChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
