+++ MahApps.Metro/Styles/Controls.ComboBox.xaml
@@ -1,8 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:Controls="clr-namespace:MahApps.Metro.Controls"
-                    xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
-                    xmlns:behaviours="clr-namespace:MahApps.Metro.Behaviours">
+                    xmlns:Controls="clr-namespace:MahApps.Metro.Controls">
 
     <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter" />
 
@@ -140,11 +138,8 @@
                                                             FontFamily="Marlett"
                                                             Content="r"
                                                             Visibility="{Binding RelativeSource={RelativeSource AncestorType={x:Type ComboBox}}, Path=(Controls:TextBoxHelper.ClearTextButton), Converter={StaticResource BooleanToVisibilityConverter}}"
-                                                            IsTabStop="False">
-                                                        <i:Interaction.Behaviors>
-                                                            <behaviours:ClearTextButtonBehavior />
-                                                        </i:Interaction.Behaviors>
-                                                    </Button>
+                                                            Controls:TextBoxHelper.IsClearTextButtonBehaviorEnabled="{Binding RelativeSource={RelativeSource AncestorType={x:Type ComboBox}}, Path=(Controls:TextBoxHelper.ClearTextButton), Mode=OneWay}"
+                                                            IsTabStop="False" />
                                                     <Rectangle x:Name="BtnArrowBackground"
                                                                Grid.Column="2"
                                                                Fill="Transparent"
