+++ MahApps.Metro/Styles/Controls.PasswordBox.xaml
@@ -1,9 +1,7 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:Controls="clr-namespace:MahApps.Metro.Controls"
-                    xmlns:Converters="clr-namespace:MahApps.Metro.Converters"
-                    xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
-                    xmlns:behaviours="clr-namespace:MahApps.Metro.Behaviours">
+                    xmlns:Converters="clr-namespace:MahApps.Metro.Converters">
 
     <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="pack://application:,,,/MahApps.Metro;component/Styles/Shared.xaml" />
@@ -174,11 +172,8 @@
                                     Content="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ButtonContent), Mode=OneWay}"
                                     Visibility="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ClearTextButton), Mode=OneWay, Converter={StaticResource BooleanToVisibilityConverter}}"
                                     Width="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:ControlsHelper.ButtonWidth), Mode=OneWay}"
-                                    IsTabStop="False">
-                                <i:Interaction.Behaviors>
-                                    <behaviours:ClearTextButtonBehavior />
-                                </i:Interaction.Behaviors>
-                            </Button>
+                                    Controls:TextBoxHelper.IsClearTextButtonBehaviorEnabled="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ClearTextButton), Mode=OneWay}"
+                                    IsTabStop="False" />
                             <ContentPresenter x:Name="PART_CapsLockIndicator"
                                               Grid.Row="0" 
                                               Grid.Column="1"
@@ -432,11 +427,8 @@
                                     FontFamily="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ButtonFontFamily), Mode=OneWay}"
                                     Content="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ButtonContent), Mode=OneWay}"
                                     Template="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:TextBoxHelper.ButtonTemplate), Mode=OneWay}"
-                                    IsTabStop="False">
-                                <i:Interaction.Behaviors>
-                                    <behaviours:ClearTextButtonBehavior />
-                                </i:Interaction.Behaviors>
-                            </Button>
+                                    Controls:TextBoxHelper.IsClearTextButtonBehaviorEnabled="True"
+                                    IsTabStop="False" />
                             <ContentPresenter x:Name="PART_CapsLockIndicator"
                                               Grid.Row="0"
                                               Grid.Column="1"
